2.0.0
Evolution started. Crasher optimized

+++ Assets/Scripts/Power.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class Power : MonoBehaviour
 {
-    public int power = 1;
+    public float power = 1;
 
-    public int maxPow = 5;
+    public float maxPow = 5;
     // Start is called before the first frame update
     void Start()
     {
-        power = PlayerPrefs.GetInt("Power");
+        power = PlayerPrefs.GetFloat("Power");
     }
 
     // Update is called once per frame
